requirements.txt
Django==4.2.7
django-multiselectfield==0.1.12
